requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-groq>=0.1.0
langgraph>=0.0.10
networkx>=3.2.1
matplotlib>=3.8.2
python-dotenv>=1.0.0
docstring-parser>=0.15
requests>=2.31.0
outlines>=0.1.1
docling>=1.16.0
transformers>=4.37.2
pydantic>=2.5.3
streamlit>=1.32.0
pyvis>=0.3.2
numpy>=1.24.0
scipy>=1.11.0
sentence-transformers>=2.5.1
python-louvain
